+++ samples/node/app.js
@@ -10,7 +10,8 @@ var logger = residue.getLogger('sample-app');
 
 var app = express();
 
-var residueParams = {
+/*
+const residueParams = {
     url: "localhost:8777",
     access_codes: [
         {
@@ -18,7 +19,7 @@ var residueParams = {
             code: "a2dcb"
         }
     ],
-    application_id: "com.muflihun.securitybox",
+    application_id: "com.muflihun.securitybox",    
     rsa_key_size: 2048,
     utc_time: true,
     time_offset: 0,
@@ -26,20 +27,30 @@ var residueParams = {
     client_private_key: "/Users/majid.khan/Projects/residue/samples/clients/netcat/client-256-private.pem",
     server_public_key: "/Users/majid.khan/Projects/residue/samples/clients/netcat/server-1024-public.pem"
 };
+*/
 
 app.get('*', function(req, res, next) {
     logger.info('Request: ' + req.url);
     return next();
 });
 
+function namedFunc() {
+    logger.debug('this is from named func');
+}
+
 app.get('/', function (req, res) {
     logger.info('Another log');
-    logger.info('Info Hello World!');
+    logger.debug('Info Hello World!');
+    namedFunc();
     res.send('Hello World!');
 });
 
 app.listen(3009, function () {
     console.log('Open http://localhost:3009 on browser');
 
-    residue.connect(residueParams);
+    // Either you can use residueParams or loadConfiguration form file
+    // for this sample we use loadConfiguration
+    if (residue.loadConfiguration('client.conf.json')) {
+        residue.connect(/*residueParams*/);
+    }
 });