+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## [Unreleased]
 ### Fixed
 - Fixed log for generating key size
+- Fixed issues with getting correct filename, line number and function name
+
+### Added
+- Added `loadConfiguration` to load configurations from json file
 
 ## [1.0.0-beta.6]
 ### Fixed